resolve circular imports
