+++ package.json
@@ -31,6 +31,7 @@
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "rollup": "^4.12.0",
+        "rollup-plugin-auto-external": "^2.0.0",
         "rollup-plugin-delete": "^2.0.0",
         "rollup-plugin-dts": "^6.1.0",
         "rollup-plugin-peer-deps-external": "^2.2.4",
