+++ src/components/Table/utils.tsx
@@ -1,15 +1,3 @@
-import React from "react";
-
-import {
-    BooleanCheckAccessor,
-    ColoredTextAccessor,
-    ColumnAccessorType,
-    DefaultTextAccessor,
-    SparkPercentageBarAccessor,
-} from "@table/index";
-
-import { resolveNAs } from "@renderers/utils";
-
 export const resolveJSONFieldValue = (value: string) => {
     try {
         const jsonValue = JSON.parse(value);
@@ -66,42 +54,5 @@ export const parseFieldValue = (value: any, nullStr: string = ""): any => {
             throw new Error(
                 `ERROR: Unable to parse field value - unhandled ColumnAccessor type: ${value.type.name}`
             );
-    }
-};
-
-export const resolveColumnAccessor = (
-    key: string,
-    accessorType: ColumnAccessorType | any = "Default"
-) => {
-    const MemoSparkPercentageBarAccessor = React.memo(SparkPercentageBarAccessor);
-    const MemoBooleanCheckAccessor = React.memo(BooleanCheckAccessor);
-    const MemoColoredTextAccessor = React.memo(ColoredTextAccessor);
-    const MemoDefaultTextAccessor = React.memo(DefaultTextAccessor);
-
-    switch (accessorType) {
-        case "PercentageBar":
-            return (row: any) =>
-                resolveNAs(
-                    row[key],
-                    <MemoSparkPercentageBarAccessor
-                        value={{ value: row[key], percentage: row[key] * 100 }}
-                    />
-                );
-        case "BooleanCheck":
-            return (row: any) => <MemoBooleanCheckAccessor value={row[key]} />;
-        case "ColoredText":
-            return (row: any) =>
-                resolveNAs(
-                    row[key],
-                    <MemoColoredTextAccessor value={row[key]} htmlColor="red" />
-                );
-        case "Link":
-            return (row: any) =>
-                resolveNAs(row[key], <MemoDefaultTextAccessor value={row[key]} />);
-        case "ScientificNotation":
-        case "Float":
-        default:
-            return (row: any) =>
-                resolveNAs(row[key], <MemoDefaultTextAccessor value={row[key]} />);
     }
 };
