+++ src/components/Text/BooleanCheck.tsx
@@ -3,7 +3,7 @@ import React from "react"
 import Box from "@mui/material/Box"
 import CheckIcon from "@mui/icons-material/Check"
 
-import { TextRenderer } from "@renderers/types"
+import { TextRenderer } from "@text/types"
 
 export const BooleanCheck: React.FC<TextRenderer> = ({ value, className, htmlColor, muiColor }) => {
     if (value && ["true", "yes"].includes(value.toString().toLowerCase())) {
