+++ src/components/Text/NASpan.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { TextRenderer } from "@renderers/types";
+import { TextRenderer } from "@text/types";
 
 export const NASpan: React.FC<TextRenderer> = ({value="N/A", className="grey"}) => {
     return (
